backend/requirements.txt
astroid==2.15.4
Click==8.1.3
ecdsa==0.18.0
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
future==0.18.3
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
pycryptodome==3.17
pylint==2.17.3
python-jose-cryptodome==1.3.2
six==1.16.0
typed-ast==1.5.4
Werkzeug==2.3.2
wrapt==1.15.0
Flask-Cors==3.0.10
